refactor: Refactoring Swiper to common component
